code reuse. fix http_client error.
