+++ package.xml
@@ -103,6 +103,7 @@
 			<file role="src" name="swoole_buffer.c" />
 			<file role="src" name="swoole_http_server.c" />
 			<file role="src" name="swoole_http_v2_server.c" />
+			<file role="src" name="swoole_http_client.h" />
 			<file role="src" name="swoole_http_client.c" />
 			<file role="src" name="swoole_http.h" />
 			<file role="src" name="swoole_websocket_server.c" />
